Logs are now clear endpoints
